Rename repo, remove deprecated Quarto doc, and update README (#141)
## Background

This PR makes two main changes to the codebase:

1. Renames all references to `pinval` ➡️ `homeval` to match the recent
renaming of the repo
* Note that we exclude AWS resources from this change, since it would be
too much of a pain to recreate those, and they're largely internal so
it's not as big of a deal
2. Removes the deprecated Quarto doc and its associated renv environment
for the sake of clarity

We also take a quick pass at the README to make sure the development
instructions are up-to-date, and add a new section on deployment.

## Testing

I tested this locally on a random PIN to confirm the script continues to
work:

```
AWS_ATHENA_S3_STAGING_DIR="s3://ccao-athena-results-us-east-1" python3 scripts/generate_homeval/generate_homeval.py --run-id 2025-09-07-adoring-jean --pin 20113180020000 --skip-html
```

Note that we are not able to test on CI, since the `generate-homeval`
workflow will not be available in the Actions tab of this repo until we
merge this PR to the main branch. I'll plan to do a quick test on prod
after merging and will put up a fast follow PR to fix any bugs that
shake out in the process. This is not an ideal deployment workflow, but
since we're not actively developing or deploying this app right now, I
figure that the stakes are not very high.

## Migration

A few changes to note once you pull these changes in:

* Any existing reports in the `hugo/content/pinval-reports/`
subdirectory will not be automatically moved to
`hugo/content/homeval-reports/`, so if you want to preserve those
reports you should rename the subdirectory yourself
* If you have a Python virtual environment at
`scripts/generate_pinval/.venv/`, you will have to manually move it to
`scripts/generate_homeval/`
* It may be easiest to just delete and recreate the virtual environment,
since this PR also renames the Python project