Update `ccao` package source (#150)
Now that the `ccao` Python package [is deployed to
PyPI](https://pypi.org/project/ccao/1.3.1/)
(https://github.com/ccao-data/ccao/pull/44), this PR updates the package
source in the uv lockfile for the `generate_homeval.py` script in this
repo.

We also take this opportunity to fold in another small fix that I
noticed while testing: The example comps run ID in the
`generate_homeval.py` docstring is currently out of date, which leads to
an esoteric error if you try to run the example commands. This PR
updates those example commands to use the correct comps run ID for 2025.

I ran a [test
workflow](https://github.com/ccao-data/homeval/actions/runs/19510548145/job/55848907599)
and confirmed that it still works as expected using the new dependency.