Consolidate significant sales qc analysis doc (#151)
This modifies the existing
[sales_qc](https://github.com/ccao-data/enterprise-intelligence/blob/main/ei-issue-0050/ei-issue-0050.qmd)
doc to compare multiple runs. It is currently set to use 2025 run id's,
but that can be modified to use the 2024 runs by uncommenting out the
params.

It groups the output table by run_id and creates tabsets for the charts
of the characteristics. These are modified to use the same sample of
2000 pins for each run_id.